Moved tests over to esm
